Port build to meson+pyproject
To work with python >= 3.12, which will deprecated numpy.distutils, port
the build to meson and pyproject.

    $ pip install build
    $ python -m build
    $ cd test && python -c "import fftlog; print(fftlog.__all__)"
    ['fhti', 'fftl', 'fht', 'fhtq']

The cd test is to avoid trying to import from $(pwd)/fftlog.

This just changes the build, and does not hard constrain versions.
Follow up work should be to figure out true minimum versions, rework CI
and deploy, and maybe test with different compilers, and fix the type
error in ifac and friends.