Port build to meson+pyproject (#8)
To work with python >= 3.12, which will deprecated numpy.distutils, port
the build to meson and pyproject.

    $ pip install build
    $ python -m build
    $ cd test && python -c "import fftlog; print(fftlog.__all__)"
    ['fhti', 'fftl', 'fht', 'fhtq']

The cd test is to avoid trying to import from $(pwd)/fftlog.

This just changes the build, and does not hard constrain versions.
Follow up work should be to figure out true minimum versions, maybe test
with different compilers, and fix the type error in ifac and friends.

The directory tree is flattened because fftlog/ doesn't form a cohesive
module as-is. Due to the presence of the __init__.py the dir would be
detected as a module and loaded by pytest and friends, shadowing the
installed one and failing because it does not find the native extension.

Finally, remove __init__.py from tests/ to not pretend it is a module
anymore.